backend/requirements.txt
asgiref==3.11.0
django-anymail[sendgrid]==12.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
channels==4.2.0
channels-redis==4.2.0
daphne==4.1.2
stripe==11.3.0
PyJWT==2.8.0
cryptography==41.0.7
gunicorn==25.1.0
dj-database-url==2.1.0
whitenoise==6.9.0
uvicorn==0.34.0
